+++ changelog.md
@@ -1,9 +1,13 @@
+### 0.1.6 - 2018-09-28
+- property `actionInitiated` is now of type `number` which is incremented when the `create`, `close` 
+action on an entity is under process and decremented when the action completes (succeeeded or failed).
+
 ### 0.1.5 - 2018-09-27
 - Improved log statements for better debugging.
 - Any type of `error` event will be emitted with a tick delay. This would give enough time for the 
 `create()` methods to resolve the promise.
-- Added a new `number` property `actionInitiated` which is incremented when the `create`, `close` 
-action on an entity is under process and decremented when the action completes (succeeeded or failed).
+- Added a new `boolean` property `actionInitiated` which indicates whether the `create`, `close`
+action on an entity is under process.
 
 ### 0.1.4 - 2018-09-25
 - `options` is a required property of `Connection` and `Container`.